pyproject.toml
[project]
name = "travel-skill"
version = "0.1.0"
description = "Unified MCP server for travel: flights, hotels, Airbnb, and public transit"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "flights>=0.7.0",
    "selectolax>=0.3.0",
    "requests>=2.31.0",
    "playwright>=1.40.0",
    "primp>=0.6.0",
]

[project.scripts]
travel-mcp = "travel_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/travel_mcp"]
